Break waveform/ ↔ library/ and analyzer/ ↔ library/ include cycles
Part of a broader effort to clean up include dependencies as a step
towards a modular build.

Extracted AnalysisType (unscoped enum) and AnalysisInfo struct from
AnalysisDao into a new thin header library/dao/analysisinfo.h.
AnalysisInfo requires only track/trackid.h and Qt types — no waveform
or DAO machinery.

AnalysisDao now includes analysisinfo.h and re-exports the types via
using aliases (AnalysisDao::AnalysisType, AnalysisDao::AnalysisInfo) and
static constexpr aliases for the enum values (TYPE_WAVEFORM etc.) so
that all existing call sites continue to compile unchanged.

Removed include library/dao/analysisdao.h from waveformfactory.h,
replacing it with library/dao/analysisinfo.h. Updated parameter type
AnalysisDao::AnalysisInfo → AnalysisInfo in the declaration and
definition of WaveformFactory::loadWaveformFromAnalysis.
This removes the waveform/ → library/ include dependency from waveformfactory.h.

Changed AnalyzerWaveform::m_analysisDao from a by-value AnalysisDao
member to std::unique_ptr<AnalysisDao>. The header now forward-declares
AnalysisDao and includes <memory>; the full AnalysisDao definition is
pulled in only by analyzerwaveform.cpp. Updated all m_analysisDao.
call sites to m_analysisDao->. Constructor initializer updated to
std::make_unique<AnalysisDao>(pConfig).
This removes the analyzer/ → library/ include dependency from analyzerwaveform.h.
